samsung-ril: fix client.c compile
ril_clients is actually an array of struct.  In files like client.c and samsung-ril.c, it follows behind a for-loop statement which checks for the count of ril_clients_count.
Thus, we can safely ommit this line as a similar approached is observed after this faulty line:

    hardware/ril/samsung-ril/client.c:32:6: error: comparison of array 'ril_clients' equal to a null pointer is always false [-Werror,-Wtautological-pointer-compare]
            if (ril_clients == NULL || ril_clients_count == 0)
                ^~~~~~~~~~~    ~~~~

For best practices, the check for 'ril_clients_count == 0' is being retained.

Change-Id: I581b13b5e11d58efbc7636bd7181d40482fd696f